calculation of disease incidence

This function calculates the frequency of the disease in the observed population, i.e. the number of sampled units (leaves, bunches, plants, etc.) infected expressed as a percentage of the total number of units observed.

Utilisation

incidence(vecteur)

Arguments

vecteur

a numeric vector of disease severity observations in percentage

Valeur de retour

a numeric value of incidence, in percent

Exemples

vec <- c(1, 2, 3, 0, 5, 6, 0)
incidence(vec)
#> [1] 71.42857