Barplot summary for an experiment

This function automatically generates a bar graph from observation or experimental data.

Utilisation

plot_xpbar2(
  data2plot,
  xcol = "xp_trt_code",
  ycol = "value",
  fillcol = NULL,
  scale = NULL,
  bar_color = NULL,
  bar_width = 0.7,
  border_tnt = TRUE,
  show_errorbar = TRUE,
  ...
)

Arguments

data2plot

a dataframe, the first col is used for x axis, col value for y axis and col type for fill.

xcol

col to be used as factor for x avis

ycol

col to be used for y axis

fillcol

col to be used for fill

scale

Upper limit of the y axis. If NULL, automatic adjustment.

bar_color

Bar fill colour.

bar_width

bar width

border_tnt

if true, the border color is set as red for TNT

...

other parameters for labs (title, x, y,fill)

Valeur de retour

a barplot

Exemples

# Exemple d'utilisation
plot_xpbar(my_data)
#> Error: objet 'my_data' introuvable