Plot a Heatmap of Experimental Data

Generates a heatmap based on experimental observation data, using plot_x and plot_y coordinates and coloring according to a selected disease variable (_PC columns).

Utilisation

plot_xpheat(
  data,
  variable,
  titre = NULL,
  echelle = NULL,
  orientation = c("vertical", "horizontal"),
  residus = FALSE,
  fill = NULL,
  caption = NULL
)

Arguments

data

A dataframe containing at least 'plot_id', 'plot_x', 'plot_y', the variable of interest, and optionally 'calculation'.

variable

Character. The name of the column representing the variable to display (e.g., "PM_LEAF_PC").

titre

(optional) Character. Title of the plot. If NULL, a default title based on the variable is used.

echelle

(optional) Numeric. Maximum value for the fill color scale. If NULL, an automatic scale is calculated.

orientation

Character. Either "vertical" or "horizontal" to control plot orientation. Default is "vertical".

residus

Logical. If TRUE, centers the values by subtracting the global mean (residual visualization).

fill

Optional. A vector of two colors (e.g., c("yellow", "red")) to manually define the fill gradient. Ignored if residus = TRUE.

caption

Character. A string used as the caption text displayed at the bottom of the plot. Default is "IFV+".

Valeur de retour

A ggplot2 heatmap object.

Détails

Aggregates values by plot_id, taking the mean when multiple observations exist.

If no *_PC variable is found in the data, the function will stop with an error.

Useful for visualizing spatial patterns of disease in experimental trials.