Plot a Heatmap of Experimental Data

Generates a heatmap based on experimental observation data, using plot_x and plot_y coordinates and coloring according to a selected disease variable (_PC columns).

Utilisation

plot_xpheat(
  self,
  stats = NULL,
  calculation_choices = NULL,
  max_y = NULL,
  flip_orientation = FALSE,
  yx_ratio = 0.3,
  resids = FALSE,
  midpoint = NULL,
  short_names = FALSE,
  ...
)

Arguments

self

An instance of the UserData R6 class

stats

character, the name of a stats list inside self$stats.

calculation_choices

vector of character, choice(s) of calculation to plot. For example c("FA UN_BER_PC","IA UN_BER_PC")

max_y

(optional) Numeric. Maximum value for the fill color scale. If NULL, an automatic scale is calculated.

flip_orientation

TRUE or FALSE to control plot orientation. Default is FALSE. see ?ggplot2::coord_flip()

yx_ratio

specified ratio between the physical representation of data units on the axes. see ?ggplot2::coord_fixed()

resids

Logical. If TRUE, residuals are calculated from the linear model and plotted

midpoint

numeric, value of midpoint for fill_scale. Automatically set to 0 if resids = T or more than one calculation

short_names

to have shorter names for fill removing the last part of the string after the last space occurence

...

other parameters for labs (title, x, y,fill, caption). see ?ggplot2::labs()

Valeur de retour

A ggplot2 heatmap object.

Détails

Aggregates values by plot_id, taking the mean when multiple observations exist.

Useful for visualizing spatial patterns of disease in experimental trials.

Exemples

# example code
if (FALSE) { # \dontrun{

# Before creating the graph, it is necessary to have compiled the statistics.
prepare_data(mydata, df = "data_G1")
test_stats(mydata, prep_data = "prp1_data_G1")

# 3. Generating the heatmap
plot_xpheat(mydata, stats = "prp1_data_G1")

# Example with residual calculation and selection of specific variables
plot_xpheat(mydata,
            stats = "prp1_data_G1",
            calculation_choices = c("FA UN_BER_PC", "IA UN_BER_PC"),
            resids = TRUE,
            title = "Carte d'intensité des maladies (Résidus)")
} # }